Initial commit
This commit sets up the Halogen Hooks project and demonstrates a
tiny proof of concept: a button component opting in to two state
cells in Halogen using a custom Hooks DSL.

The basic idea: Hooks are a React feature which allow users to take
their function components and selectively opt-in to React features
like state and lifecycle methods. For example, a user can write:

  ```js
  [state, setState] = useState({ myString: "" })

  return (
    <div>{state.myString}</div>
  );
  ```

to get a reference to component state and a function to update the
state.

Hook components are separate from React's class components and
enable easier reuse for stateful logic. Users can write their own
hooks to use in custom components which serve as an alternative
to higher-order components or render props.

For example, you could write a `useMousePosition` hook which
is built on top of `useState` and subscribes to and stores the
coordinates of the mouse:

  ```js
  { x, y } = useMousePosition()

  return (
    <div>The mouse is at position ({x}, {y})</div>
  );
  ```

This commit implements a prototype of the same idea for Halogen.
The core idea is to:

  - Provide a user with a DSL for writing hook components
  - Provide a usual Halogen component which can translate that
    DSL into ordinary Halogen code and execute it.

In the current implementation, a user can write a hook component:

  ```purs
  hook :: forall m. Hook String (H.ComponentHTML Action () m)
  hook = do
    Tuple helloState modifyHello <- Hook.useState 0 "Hello"
    Tuple worldState modifyWorld <- Hook.useState 1 "World"

    pure $
      HH.button
        [ HE.onClick \_ -> Just $ modifyWorld (_ <> "!") ]
        [ HH.text $ helloState <> " " <> worldState ]
  ```

and provide it to the underlying `hookComponent` for translation,
for example as the root component of an application:

  ```purs
  main :: Effect Unit
  main = runHalogenAff do
    body <- awaitBody
    runUI Hook.hookComponent { hook } body
  ```