Remove hook identifiers and add indexed free monad
Previously you needed to provide a persistent, unique identifier
to a hook (ie. Hook.useState <id> \_ -> initialState). That was
necessary because there had to be a way to associate hooks with
their state cells.

This was extremely brittle, however.

  - There was no way to verify your hook did use unique identifiers
    and no hooks shared the same identifier.

  - It was easy to lose track of whether an identifier was used
    if you combined several hooks into one component, especially
    if those hooks themselves under the hood had their own unique
    identifiers.

This commit switches the representation of the Hook DSL to use
an indexed free monad instead of a regular free monad. That
change ensures that the binds used to produce a hook are always
run in the same  order. In other words, it's no longer possible
to do something like this:

```purs
state /\ setState <-
  if true then
    _ <- useState \_ -> 10
    useState \_ -> ""
  else
    _ <- useState \_ -> ""
    useState \_ -> 10
```

Since we can now ensure that hooks are always evaluated in the
same order we can remove the identifiers from the user's view
as well.

When the hook DSL is run for the first time (initialization),
each bind gets its own cell in a state array according to its
position in the list of binds. On subsequent runs we cycle
through the state array, at each step associating a particular
hook with a particular index in state.

This is safe to do because the order of hooks cannot change and
a hook cannot be removed from the tree. In other words, the
relation from hook -> index is stable.