Support providing input to a hook component
This commit enables using Halogen's input / receiver mechanisms
to support a stream of input values into a hook-based component.

Instead of providing a hook to the hook component function you now
provide a function from (input -> hook). That way we can store the
input in state and provide it to the hook function any time it
changes.

Input in hook components therefore behaves quite similarly to how
props in React work -- they are kept logically separate from the
component state and you can always reference them (read-only).

The standard Halogen approach is to copy relevant parts of input
into state (sometimes called "derived state"), which lets you
skip a re-render if you only care about part of the input.

Given that the focus for now is on handling basic functionality
like supporting input in the first place I haven't worried about
performance. However, in the worst case a higher-order component
can be used to memoize the input passed to the hook component.

See:
https://reactjs.org/docs/hooks-faq.html#how-do-i-implement-shouldcomponentupdate