Introduce EvalHookM monad to expand the core Halogen feature set
Prior to this change:

  - Users used hooks from `HookF` to opt-in to Halogen features
    and provided a hook which returned `ComponentHTML (Hook.Action)`
    in order to create a component.

  - Each hook from the library returned a `Hook.Action` action which
    the underlying component implemented a handler for, which itself
    ran in `HalogenM` with access to all the state cells.

  - The returned hook action baked in the unique identifier for the
    state cell created to support it.

This led to two undesirable consequences.

First, because the id for a state cell was baked in to the returned
`Hook.Action` it was associated with, you couldn't mix together
different states and actions. This situation was impossible:

```purs
data Action = Add1

...

stateA /\ modifyStateA <- useState 0

stateB /\ actionB <- useAction "" case _ of
  Add1 -> modifyStateA (_ + 1)
```

both because `modifyStateA` wasn't a function but rather a sum
type and because the second hook, `useAction`, only was able to
modify the string state it was initialized with.

Second, the implementation of `handleAction` within the underlying
component was needlessly complex with all sorts of `mapAction` and
`imapState` calls going on.

Both of these issues have been addressed by a single major change:
the introduction of a new `EvalHook` DSL and corresponding `EvalHookM`
monad. This DSL is a replacement for `HalogenM` which will eventually
supports the same features as HalogenM as well as a few tweaks and
modifications specific for hooks.

The biggest modification is to how state is handled. `EvalHookM` no
longer tracks a single state type; instead, it requires that you
provide a 'state token' when attempting to access or modify state.

`useState` now returns this token, ensuring that the token is only
compatible with the state you provided as your initial state. The
token can then be used in calls to `modify_`, `put`, `get`, and so
on. The only way to get a state token is to use a hook, and the
only way to access state is with the token.

With this change, the earlier example could be written:

```purs
stateA /\ tokenA <- useState 0
stateB /\ tokenB <- useState ""

let
  handleClick :: forall output m. EvalHookM output m Unit
  handleClick = do
    modify_ tokenA (_ + 1)
    modify_ tokenB (_ <> "b")
```

where you can freely mix calls to different states together in a
handler very similar to the typical Halogen `handleAction`.