Add the first custom hook example
The first hook to make it into the examples is the `useWindowWidth`
hook, which saves the window width in state, subscribes to resize
events on the window, and updates the saved width each time it
changes. This logic is usually hard to share in Halogen because it
is tied in to component lifecycles and state.

Ordinarily you’d pull all the relevant code out into helper
functions — a function to subscribe, a function to unsubscribe,
perhaps a lens to update the component state at a particular field
name — and you’d use them in the initializer, finalizer, and then
some action which gets fired every time the subscription updates.

This can lead to a lot of boilerplate across many components. The
next logical step is to bundle it into a higher-order or renderless
component. But these are complex: higher-order components require
proxying inputs, outputs and renderless components force you to
extend an existing component instead of writing your own exactly
how you’d like. Both are heavyweight solutions.

Hooks are different. With Hooks you can write all the stateful logic
— including lifecycles — into a hook function, which can then be used
to entirely manage the window width in any component that uses it.
The only requirement is that the hook is used in a hook-based component.

This commit also adjusts the examples to match the Halogen examples
repository structure.