Add the useMemo hook
It's a best practice to keep computations separate from the render
function in an ordinary Halogen component -- after all, you don't
want to re-compute a pontentially expensive value every render in
your app. Traditionally you would do that by caching in state
anything you don't want to compute anew each render.

In the Hooks model, the entire Hooks.do block is re-run on every
render. The same idea applies: you don't want to recompute anything
expensive in that block if you can possibly avoid it.

You can use the same solution as is used in an ordinary Halogen
component: cache expensive-to-compute values in state.

This commit introduces a different solution: use the `useMemo` hook
to cache a value and only recompute it when dependencies change.
If you usually declare a value like this:

```purs
Hooks.component \input -> Hooks.do
  let x = input.count + 5
  ...
```

it will recompute `input.count + 5` on every render, even when the
input has not changed (same as the old Halogen would, though you
used to *have* to copy inputs to state). To avoid that, we can use
useMemo:

```purs
Hooks.component \{ count } -> Hooks.do
  x <- Hooks.captures { count } $ flip Hooks.useMemo \_ -> count + 5
  ...
```

Now, the `count + 5` computation will only re-run and recompute `x`
if the value of `count` in the input has changed.