Add UseRef hook and fix mutual recursion in interpreters
This commit introduces two major changes.

First, it introduces a new UseRef hook which allows users to create
a mutable reference and then refer to its current value within the
scope of a hook. This was necessary because while you could easily
store a reference in state (as usual) you couldn't refer to the
value within without an effect to read the ref. There are plenty of
use cases for easy access to an isolated piece of mutable state, and
this hook makes that possible.

```purs
Hooks.do
  value /\ ref <- Hook.useRef 0
```

In the above block of code, `value` is a value of type `Int` which
will be updated anytime the hook is re-run, and `ref` is a value of
type `Ref Int` which you can use to mutate the reference at any time.

References like these can be used to improve performance: when you
have a piece of state which changes frequently but you don't want
to store in actual state (because it will trigger a re-render each
state change) you can use a ref instead. For example you might store
a scroll position and have it update each time the user scrolls, but
without causing a re-render until you actually trigger an event, such
as with a debouncer that /does/ update state.

The other major change in this commit introduces a queue for running
the hook interpreter. Previously, if the hook interpreter ran a user
action which needed to interpret the hooks again (for example, the
user's action modifies state multiple times) then the calls would
be processed recursively. All state updates would work just fine,
but once the recursive calls resolved the resulting HTML would be
that of the /first/ interpreter. That meant current state and stale
HTML and in general a buggy implementation.

This change instead places those recursive calls into a queue. When
each interpretation completes the next item in the queue is processed
until they all complete and a new action causes the full cascade
all over again.

Suggestion originally from @natefaubion.