Add useDebouncer custom hook example
Debouncing was actually the thing which got me excited about hooks.
The debouncer in libraries like purescript-halogen-select and
purescript-halogen-formless are complicated and weave their way
through the actual component implementation. It would be a whole
lot nicer to avoid this with a simple useDebouncer hook which could
be shared with other libraries.

This commit adds the final custom hooks example to this repository:
the debouncer that motivated this in the first place.