Move internal hook state to a mutable reference
Halogen is designed to re-render after each call to update state;
if the state is unchanged then referential equality will prevent
a render from occurring (this is why `get` is fine, and why `put`
on an identical state does not force a re-render).

With hooks, though, there's a lot of state held in Halogen state
that needs to be updated without causing a re-render (for performance
reasons). In fact, the only time a re-render should occur is when
new HTML has been produced (or else there's no point in re-rendering
anyway).

With that in mind the only field that should exist in the state and
not behind a `Ref` is the `html` field.

This commit moves all other state into a mutable reference which
is updated when ordinary state updates would have occurred.
Conceptually it's the same as the same state updates which occurred
before -- it's just as if those state updates were gated to not
trigger a re-render.

For that reason I don't consider the use of a hidden, non-user
-accessible mutable state cell to be a good reason to propagate a
`MonadEffect` constraint all the way up. It's pretty much the same,
from a user point of view, as just having the regular opaque state
inside of Halogen. For this reason I use unsafePerformEffect.

All of this creates a significant performance improvement in the
library. In testing I've seen no issues from the change, only an
improvement. I don't foresee issues but if any weird states arise
in the future the reference may be the first place to look.