Rename HalogenHookM -> HookM and HookF -> UseHookF
`HookF` is the first type introduced to this library and it defines
the Hook API. But users didn't see this type; instead, they would
usually use `HalogenHookM`. It seemed a shame to use the long form
everywhere when users never got to use the short form, so I have
updated the names so that `HookF` becomes `UseHookF` an
`HalogenHookM` becomes `HookM`.

As part of the change I renamed the `coerce` function to `publish`
in response to feedback and used the long form of type variables
for `slots` and `output` in the examples section.