Support opaque custom hooks as newtypes
While it is possible to define hook types for your custom hooks as
simple type synonyms, this is not advised for a few reasons:

  1. Error messages suffer: if you use several custom hooks, each
     of which is built atop primitive hooks like UseState or other
     custom hooks, and they're all defined as type synonyms, then
     an improper usage will spit out an explosion of expanded types
     as the error.

  2. Any internal types you used to define the hooks in the type
     synonym also have to be exported. For example -- you have to
     export your internal state type if you use UseState, which is
     against the principle that hook consumers shouldn't care about
     their internals.

However, defining custom hook types was previously a little funky.

  1. You'd first define a type synonym specifying the hooks you're
     using in your implementation named UseHook'

  2. Then you'd foreign import a new data type of the same kind as
     your type synonym (usually Type -> Type) named UseHook. This
     is what you'd export from the module.

  3. Then you'd write the internal implementation of your hook, of
     the type `Hook ps o m UseHook' a`. This is not exported.

  4. Your internal implementation would go in a `where` clause, and
     the actual `useHook` definition would just be:

         useHook :: Hook ... UseHook ...
         useHook = Hooks.coerce hook
           where
           hook :: Hook ... UseHook' ...
           hook = Hooks.do ...

     which was necessary so you could ensure that you still got
     compile-time checks on your Hook implementation while coercing
     the result to your opaque tag.

This is all a bit much, especially having to define the internal
implementation locally with an explicit type signature to prevent
you from inadvertently coercing something incorrect into your tag.

The new approach uses simple newtypes. You now define your custom
hook type as a newtype ala

    newtype UseHook h = UseHook (UseEffect h)
    derive instance newtypeUseHook :: Newtype (UseHook h) _

which can be exported from your module. Then, when you implement
your custom hook you give it this type and use the `Hooks.wrap`
function:

    useHook :: Hook ps o m UseHook a
    useHook = Hooks.wrap Hooks.do ...

The `wrap` function ensures that the hooks used in the body of the
`useHook` definition match with the hooks used in your newtype. The
body of the `useHook` definition is checked by the compiler to
ensure the hooks are used correctly.