Skip evaluation on referentially-equal states. (#21)
Hooks use the same machinery for state updates as Halogen does,
which means that calls to `get` state are just calls to `Modify`
state with the identity function.

Previously this meant that hooks would be re-run on calls to get
state. Just like with Halogen, calls to get state should skip any
evaluation via unsafeRefEq.

This also updates the test suite to verify that calls to get state
do not trigger hook evaluations.