Introduce tokens for component-only features (queries, slots, output) (#22)
One of the first challenges I faced implementing Hooks was what to do about queries. The request / response nature of queries make them only useful in the context of a component, not an arbitrary hook, and so I needed a way to restrict them so queries could only be used when you created a component.

@natefaubion came up with the idea of a 'token', provided by the component function, where you can only use the useQuery hook if you have this token via the component function. That ensures that queries are only supported in components.

As I've worked with Hooks a little longer I realize the same should have been done for output message and child slot types, too. While it's possible to use functions that operate on these types within Hooks (specifically within HookM), they really only make sense in the context of a component, and they were always a poor fit for Hooks: if you ever specialized the output or slot type then you could only ever use the hook with a component with that specific type, which makes the hook essentially unusable for shared logic.

This pull request remedies the situation by restricting output messages and child slots (as well as functions that require them: query, queryAll, and raise) to only be possible when using the component function, via the same token mechanism as queries.