Re-evaluate hooks if effects may have modified dependencies (#20)
Effects run after Hook evaluation, but without this change, if the effects happened to change the dependencies of other effects via state updates they would not trigger Hooks to evaluate again. This led to incorrect behavior where state updates during effects failed to cause other effects to run.

With this change, if any effects are queued after Hook evaluation then we assume there's a chance that dependencies will be modified, and we need to evaluate all the Hooks again after running the effects. This is less efficient (running an effect will lead to two Hook evaluations), but the behavior is correct. Only effect hooks are affected.