Support skipping evaluations when input is unchanged
When a Halogen component re-renders it sends input again to any
of its child components. When a Hook component receives new input
it will store that input and re-evaluate all Hooks.

This can be a performance issue if a Hook function (especially an
expensive one) is run unnecessarily each time the parent component
renders.

This commit introduces `memoComponent`, a version of the `component`
function which allows Hook components to ignore new input if it is
unchanged from previous input.