Return a modify function from useState instead of a token (#29)
Replaces state tokens with a modify function returned by useState. This brings the implementation more in line with other Hooks implementations like `purescript-react-hooks` and the official React Hooks implementation. It means that the Halogen state functions are no longer used.

When users need to retrieve a current state value and can't use the state value returned by `useState` (for example, in an asynchronous function), they should store the state they need in a mutable reference. This PR also adds a `useGet` example to demonstrate that.