Use mutable state internally.
Hooks has always shipped with mutable internal state -- all
internal state was always behind a `Ref` so that updating it would
not trigger renders.

However, we were storing immutable data in a mutable reference. Each
operation on that data still creates a new data structure. This
commit changes the internals to be fully mutable, which means that
Hook evaluations occur on the same mutable object internally.

Consumers of the library should see no change -- this is purely an
internal optimization.