Add ability to measure performance impact of changes (#53)
ooks carry some overhead when compared to normal Halogen components because they are a layer on top. The library right now makes essentially no attempts to optimize the overhead that Hooks carry (I made an attempt in #34 but ultimately did not merge it because I didn't have a reliable way to measure the impact). I would like to explore ways to make Hooks more performant in the future, and I'd also like to be able to test whether future changes introduce performance regressions.

This PR introduces some scaffolding to measure the performance impact of changes made to Hooks. It uses Puppeteer to run Halogen applications and measure how long they take to execute, how much memory they use, and what average FPS they maintain. These three metrics can be used to judge whether changes to the library are helping or hurting. They should not be used, however, to make an absolute judgment on the performance of Hooks in a real world application as they are minimal, isolated benchmarks.